TestVocacionalCBTIS106 - copia - copia/TestVocacionalCBTIS106 - copia - copia/requirements.txt
flask>=2.3.0
python-dotenv>=0.21.0
markupsafe>=2.1.0
Werkzeug>=2.3.0
gunicorn>=21.2.0
